requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
# Interactive selector tool (lingq_interactive.py)
# After installing, run once: playwright install chromium
playwright>=1.40
# Web UI (lingq_app.py / start.command)
flask>=3.0
# Audio transcription — lingq_audio_import.py --transcribe
# Requires ffmpeg on PATH: brew install ffmpeg  (macOS) / apt install ffmpeg  (Linux)
openai-whisper>=20231117
